fix: restore message send logic to simplify chat view and input box
