+++ instagram-ts/source/ui/components/input-box.tsx
@@ -1,22 +1,15 @@
 import React, {useState, useEffect} from 'react';
-import {Box, Text, useInput} from 'ink';
+import {Box, useInput} from 'ink';
 import TextInput from 'ink-text-input';
-import type {InstagramClient} from '../../client.js';
-import type {ChatState} from '../../types/instagram.js';
 import {
 	getFilePathSuggestions,
 	getCommandSuggestions,
 } from '../../utils/autocomplete.js';
-import {parseAndDispatchChatCommand} from '../../utils/chat-commands.js';
-import {preprocessMessage} from '../../utils/preprocess.js';
 import {AutocompleteView} from './autocomplete-view.js';
 
 type InputBoxProperties = {
+	readonly onSend: (message: string) => void;
 	readonly isDisabled?: boolean;
-	readonly client: InstagramClient;
-	readonly chatState: ChatState;
-	readonly setChatState: React.Dispatch<React.SetStateAction<ChatState>>;
-	readonly height: number;
 };
 
 type AutocompleteState = {
@@ -24,7 +17,7 @@ type AutocompleteState = {
 	readonly isActive: boolean;
 	readonly suggestions: readonly string[];
 	readonly selectedIndex: number;
-	readonly triggerIndex: number;
+	readonly triggerIndex: number; // Position where the trigger starts
 	readonly query: string;
 };
 
@@ -38,72 +31,24 @@ const initialAutocompleteState: AutocompleteState = {
 };
 
 export default function InputBox({
+	onSend,
 	isDisabled = false,
-	client,
-	chatState,
-	setChatState,
-	height,
 }: InputBoxProperties) {
 	const [message, setMessage] = useState('');
 	const [autocomplete, setAutocomplete] = useState<AutocompleteState>(
 		initialAutocompleteState,
 	);
-	const [systemMessage, setSystemMessage] = useState<string | undefined>(
-		undefined,
-	);
+
+	// By changing the key, we force the TextInput to re-mount, which resets its internal state (including cursor position after autocomplete selection)
 	const [inputKey, setInputKey] = useState(0);
 
-	// Effect to clear system messages after a delay
-	useEffect(() => {
-		if (systemMessage) {
-			const timer = setTimeout(() => {
-				setSystemMessage(undefined);
-			}, 3000); // Clear after 3 seconds
-			return () => {
-				clearTimeout(timer);
-			};
-		}
-	}, [systemMessage]);
-
-	const handleSubmit = async (text: string) => {
-		if (!text.trim() || !client || !chatState.currentThread) {
-			return;
+	const handleSubmit = (value: string) => {
+		if (value.trim()) {
+			onSend(value.trim());
+			setMessage('');
 		}
 
-		// Reset UI state first
-		setMessage('');
-		setAutocomplete(initialAutocompleteState);
-
-		const {isCommand, systemMessage: cmdSystemMessage} =
-			await parseAndDispatchChatCommand(text, {
-				client,
-				chatState,
-				setChatState,
-				height,
-			});
-
-		if (cmdSystemMessage) {
-			setSystemMessage(cmdSystemMessage);
-		}
-
-		if (isCommand) {
-			return; // Command was handled
-		}
-
-		try {
-			const processedText = await preprocessMessage(text, {
-				client,
-				threadId: chatState.currentThread.id,
-			});
-
-			if (processedText) {
-				await client.sendMessage(chatState.currentThread.id, processedText);
-			}
-		} catch (error) {
-			const errorMessage =
-				error instanceof Error ? error.message : 'Failed to send message';
-			setSystemMessage(errorMessage);
-		}
+		setAutocomplete(initialAutocompleteState); // Reset on submit
 	};
 
 	const handleAutocompleteSelection = (suggestion: string) => {
@@ -173,7 +118,7 @@ export default function InputBox({
 			const query = filePathMatch[1] ?? '';
 			const triggerIndex = filePathMatch.index + 1;
 			setAutocomplete(previous => ({
-				...previous,
+				...previous, // Preserve existing suggestions while typing
 				type: 'filePath',
 				isActive: true,
 				query,
@@ -200,7 +145,7 @@ export default function InputBox({
 						(previous.selectedIndex - 1 + previous.suggestions.length) %
 						previous.suggestions.length,
 				}));
-				return;
+				return; // Consume event
 			}
 
 			if (key.downArrow) {
@@ -209,12 +154,12 @@ export default function InputBox({
 					selectedIndex:
 						(previous.selectedIndex + 1) % previous.suggestions.length,
 				}));
-				return;
+				return; // Consume event
 			}
 
 			if (key.escape) {
 				setAutocomplete(initialAutocompleteState);
-				return;
+				return; // Consume event
 			}
 
 			if (key.tab || key.return) {
@@ -224,14 +169,14 @@ export default function InputBox({
 					handleAutocompleteSelection(selectedSuggestion);
 				}
 
-				return;
+				return; // Consume event, preventing submission
 			}
 		}
 
 		// Priority 2: Default submission on Enter
 
 		if (key.return) {
-			void handleSubmit(message);
+			handleSubmit(message);
 		}
 	});
 
@@ -252,17 +197,11 @@ export default function InputBox({
 					onSubmit={() => {}}
 				/>
 			</Box>
-			{systemMessage ? (
-				<Box marginTop={1}>
-					<Text color="yellow">{systemMessage}</Text>
-				</Box>
-			) : (
-				autocomplete.isActive && (
-					<AutocompleteView
-						suggestions={autocomplete.suggestions}
-						selectedIndex={autocomplete.selectedIndex}
-					/>
-				)
+			{autocomplete.isActive && (
+				<AutocompleteView
+					suggestions={autocomplete.suggestions}
+					selectedIndex={autocomplete.selectedIndex}
+				/>
 			)}
 		</Box>
 	);
